requirements.txt
groq>=0.4.1
python-dotenv>=1.0.0
behave>=1.2.6
pytest>=7.4.0
pytest-html>=3.2.0
allure-behave>=2.13.2
jinja2>=3.1.2
pyyaml>=6.0.1
requests>=2.31.0





